llama-cpp-cache-lab/pyproject.toml
[project]
name = "llama-cpp-cache-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-cpp-python>=0.3.17",
    "pandas>=3.0.1",
]
